Adding C++ library for Python to call C++ pointers
Working on compile.sh bash script that compiles C++ code
into .o and .so files for shared library. Also working on
genreflex code for C++ code to be added to cppyy
reflection library, so I don't have to explicitly call
cppyy.load_reflection_info('...') in Python code. It will
be better for large scale applications. But okay for now
since this is just an example on how to call C++ class in
Python code and it gets the job done