For Delete(), in main.cpp, changed to return x element deleted
Prior to this change, the bug in the code was that it
displayed to the user the index of the element that was
deleted. The issue was only in main.cpp, I passed in
index to cout instead of x. x holds the element that was
deleted since Delete(index) returns the element to x. I
made sure to pass x to cout