Added Find and Count Duplicate Elements C functions to Array Ops Lib
The first two functions FindDuplicateElements() and
CountDuplicateElements() use for loops and if condition
statement to check for duplicates. For when duplicate is
found, the duplicate numbers are printed to the screen.
The amount of times the duplicate appears per duplicate
element is printed to the screen. CountDuplicateElements2()
uses a Hash Table to hold the amount of times the duplicate
appears. Then loop through the hash table, if element is
greater than 1, print duplicate number appears x amount of
times